create material in pdf file
All the material are also available in pdfs now. Using jupyter
notebooks is not best practice. I used them because I was to
add the material to my personal website, and it can done quickly
with notebooks.

Moreover, a script has been added that allows us to generate
pdfs from the notebooks automatically